+++ Controller/CancelController.php
@@ -12,12 +12,12 @@ class CancelController extends PayumController
      */
     public function doAction(Request $request): Response
     {
-        $token = $this->getPayum()->getHttpRequestVerifier()->verify($request);
+        $token = $this->payum->getHttpRequestVerifier()->verify($request);
 
-        $gateway = $this->getPayum()->getGateway($token->getGatewayName());
+        $gateway = $this->payum->getGateway($token->getGatewayName());
         $gateway->execute(new Cancel($token));
 
-        $this->getPayum()->getHttpRequestVerifier()->invalidate($token);
+        $this->payum->getHttpRequestVerifier()->invalidate($token);
 
         return $token->getAfterUrl() ?
             $this->redirect($token->getAfterUrl()) :