+++ DependencyInjection/Compiler/BuildGatewayFactoriesPass.php
@@ -4,7 +4,6 @@ namespace Payum\Bundle\PayumBundle\DependencyInjection\Compiler;
 use Payum\Core\Exception\LogicException;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Reference;
 
 class BuildGatewayFactoriesPass implements CompilerPassInterface
 {