+++ DependencyInjection/Factory/Storage/Propel2StorageFactory.php
@@ -16,6 +16,9 @@ class Propel2StorageFactory  extends AbstractStorageFactory
         return "propel2";
     }
 
+    /**
+     * @param array<string, mixed> $config
+     */
     protected function createStorage(ContainerBuilder $container, string $modelClass, array $config): ChildDefinition
     {
         $loader = new XmlFileLoader($container, new FileLocator(__DIR__.'/../../../Resources/config/storage'));