+++ DependencyInjection/PayumExtension.php
@@ -11,7 +11,6 @@ use Payum\Core\Gateway;
 use Payum\Core\Registry\DynamicRegistry;
 use Payum\Core\Storage\CryptoStorageDecorator;
 use Sonata\AdminBundle\Admin\AbstractAdmin;
-use Symfony\Component\Config\Definition\ConfigurationInterface;
 use Symfony\Component\DependencyInjection\ChildDefinition;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
@@ -169,6 +168,9 @@ class PayumExtension extends Extension implements PrependExtensionInterface
         }
     }
 
+    /**
+     * @param array<string, mixed> $securityConfig
+     */
     protected function loadSecurity(array $securityConfig, ContainerBuilder $container): void
     {
         foreach ($securityConfig['token_storage'] as $tokenClass => $tokenStorageConfig) {
@@ -184,6 +186,9 @@ class PayumExtension extends Extension implements PrependExtensionInterface
         }
     }
 
+    /**
+     * @param array<string, mixed> $dynamicGatewaysConfig
+     */
     protected function loadDynamicGateways(array $dynamicGatewaysConfig, ContainerBuilder $container): void
     {
         $configClass = null;
@@ -279,6 +284,9 @@ class PayumExtension extends Extension implements PrependExtensionInterface
         return new MainConfiguration($this->storagesFactories);
     }
 
+    /**
+     * @param array<string, mixed> $storageConfig
+     */
     protected function findSelectedStorageFactoryNameInStorageConfig(array $storageConfig): string
     {
         foreach ($storageConfig as $name => $value) {