+++ Tests/Controller/AuthorizeControllerTest.php
@@ -5,7 +5,6 @@ use Payum\Bundle\PayumBundle\Controller\AuthorizeController;
 use Payum\Core\GatewayInterface;
 use Payum\Core\Request\Authorize;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\DependencyInjection\ServiceLocator;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 
 class AuthorizeControllerTest extends AbstractControllerTest