+++ Tests/Controller/RefundControllerTest.php
@@ -6,7 +6,6 @@ use Payum\Bundle\PayumBundle\Controller\RefundController;
 use Payum\Core\GatewayInterface;
 use Payum\Core\Request\Refund;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\DependencyInjection\ServiceLocator;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Response;