+++ Tests/Functional/Form/Type/CreditCardTypeTest.php
@@ -29,13 +29,11 @@ class CreditCardTypeTest extends WebTestCase
      */
     public function couldBeCreatedByFormFactory(): void
     {
-        if (Kernel::MAJOR_VERSION === 6) {
-            /** @var RequestStack $requestStack */
-            $requestStack = self::getContainer()->get(RequestStack::class);
-            $request = Request::createFromGlobals();
-            $request->setSession(new Session(new MockArraySessionStorage()));
-            $requestStack->push($request);
-        }
+        /** @var RequestStack $requestStack */
+        $requestStack = self::getContainer()->get(RequestStack::class);
+        $request = Request::createFromGlobals();
+        $request->setSession(new Session(new MockArraySessionStorage()));
+        $requestStack->push($request);
 
         $form = $this->formFactory->create(CreditCardType::class);
         $view = $form->createView();