+++ Tests/Functional/app/AppKernel.php
@@ -1,29 +1,41 @@
 <?php
 
-use Payum\Bundle\PayumBundle\PayumBundle;
-use Payum\Bundle\PayumBundle\Tests\Functional\app\AppKernelShared;
-use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
-use Symfony\Bundle\TwigBundle\TwigBundle;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Kernel;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\HttpKernelInterface;
 
-if (Kernel::MAJOR_VERSION === 4) {
-    class AppKernel extends AppKernelShared
+class AppKernel extends Kernel
+{
+    public function registerBundles(): iterable
     {
-        public function handle(Request $request, $type = HttpKernelInterface::MASTER_REQUEST, $catch = true): Response
-        {
-            return parent::handle($request, $type, false);
-        }
+        $bundles = array(
+            new Symfony\Bundle\FrameworkBundle\FrameworkBundle(),
+            new Symfony\Bundle\TwigBundle\TwigBundle(),
+            new Payum\Bundle\PayumBundle\PayumBundle(),
+        );
+
+        return $bundles;
     }
-} else {
-    class AppKernel extends AppKernelShared
+
+    public function getCacheDir(): string
     {
-        public function handle(Request $request, int $type = HttpKernelInterface::MAIN_REQUEST, bool $catch = true): Response
-        {
-            return parent::handle($request, $type, false);
-        }
+        return sys_get_temp_dir() . '/PayumBundle/cache';
+    }
+
+    public function getLogDir(): string
+    {
+        return sys_get_temp_dir() . '/PayumBundle/logs';
+    }
+
+    public function registerContainerConfiguration(LoaderInterface $loader): void
+    {
+        $loader->load(__DIR__ . '/config/config.yml');
+    }
+
+    public function handle(Request $request, int $type = HttpKernelInterface::MAIN_REQUEST, bool $catch = true): Response
+    {
+        return parent::handle($request, $type, false);
     }
 }